Cherry-pick: Remove workaround for SkTypeface::CreateFromName() on Android.
Cherry-pick of chromium crrev.com/23803011

Bug: 10730965
Bug: 10844878

Original description:
Remove workaround for SkTypeface::CreateFromName() on Android.

Originally on Android if SkTypeface::CreateTypeface() can't find a font
by name, it returns the default font. This had ever broke blink's font
fallback mechanism before we added a workaround not to return the font
if it's a fellback font when a font name can't be matched.

From Skia r11377 (https://chromiumcodereview.appspot.com/23601041/),
the function no more fallbacks to default font on Android, so we can
remove the special workaround and use the code in FontCacheSkia.cpp.

Not sharing FontCache::getFontDataForCharacter() because on Android
there is no interface to know if the returned family is really bold or
italic.

BUG=167557
TEST=fast/css/font-face-fallback.html

Change-Id: If36ecdb97dd38bedf90a24e496a84299ffff8347