Cherry Pick Fix for link highlighting of <img> links
Cherry pick of https://chromiumcodereview.appspot.com/24376003 minus
LayoutTests/*

Bug: 10929011

Original Description:
Fix for link highlighting of <img> links

Temporary fix for <img> inside <a>. I changed the node that gets passed
to the LinkHighlight object. Previously, when a node was tapped, we'd go
up the tree until we find a node that causes a "hand cursor" and use
that. Now, we do the same but we stop ascending up the tree if a node
isn't contained by its parent's box.

In the long term, the LinkHighlight should be changed to use the entire
hit testing region of the node, rather than the line box.

Note: This is relanding of https://codereview.chromium.org/24307004/.
The only change from that patch is in WebViewImpl.cpp
BUG=170673
Committed:
https://src.chromium.org/viewvc/blink?view=rev&revision=158232

Change-Id: Ia11323d33079552cf1b9720216cd933f369eb298