Cherry-pick: [Android WebView] Simplify processing of initial-scale=0 with viewportMetaZeroValuesQuirk
Cherry-pick of chromium http://crrev.com/25749002

Bug: 10742856

Original description:

[Android WebView] Simplify processing of initial-scale=0 with viewportMetaZeroValuesQuirk

It turns out that we can always treat zero initial-scale as 'auto'.
The previous logic has been copied from the current implementation,
but as experiments show, the resulting layout parameters are also
compatible with the existing implementation even with this simplified
logic.

BUG=282130

Change-Id: Ib2bd617ef74e70c6656d7f317aa53d3a274dda2d