Cherry-pick: [Android WebView] In non-WideViewport mode, initial-scale < 1 must be ignored
Cherry-pick of chromium http://crrev.com/25353006

Bug: 10566410

Original description:

[Android WebView] In non-WideViewport mode, initial-scale < 1 must be ignored

From my experiments with the existing implementation, it turns out that in the
case when WideViewport is turned off, values of initial-scale < 1 are ignored in
most cases. The layout width is set to device-width (scaled according to
target-densitydpi). This also includes 'auto' initial-scale.

An exception from the rule is the case when width is set to device-width,
in which case the initial-scale is respected.

BUG=303189

Change-Id: I3d7139ce0fe24e0c8f3f45e2209a9fdba0de5555