Prevent HTML form validation from blocking form submission.
On submit, form controls that perform validation will run and in
the current version of Blink call out to Chromium to show UI when
that validation fails and block the form from being submitted.

Unfortunately this UI lives in the chrome/ layer currently and so is
never shown for the WebView, leaving a form in a potentially
unsubmittable state.

Work around this by preventing validation from blocking form submission.
The old WebView did not support form validation so this is not a
regression and apps currently will be doing their own JS/server side
validation. Plans are ongoing upstream to move the code out
of chrome/ and into content/ so we will be able to show this UI in
the future and remove this workaround.

See upstream: https://code.google.com/p/chromium/issues/detail?id=293608

Bug: 10621067
Change-Id: I6abf40a0f41d3466906dd3541908b57579199636